feat(pubsub): subscription callback concurrency control
Add support for subscriptions to *schedule* multiple callbacks in
parallel.  Note that the number of callbacks that *run* in parallel
depends on the number of threads servicing the completion queue
associated with the subscription. The previous behavior (only one
callback at a time) remains possible via configuration in the
`pubsub::SubscriptionOptions`. The doxygen documentation for
`pubsub::SubscriptionOptions` has more details.