refactor: x-goog-api-client header definition
The GCS, Bigtable, and Spanner libraries needed a function returning the
x-goog-api-client header value. I refactored these functions to common.
Because in the case of GCS this was part of the public API, I left the
function there, but replaced the implementation. In hindsight I should
have done this refactoring *before* the refactoring for
`IsContextMDValid()`.