fix(storage): work with unknown SSL version in curl (#5037)
Sometimes libcurl will report the SSL version as `NULL`, typically
because there are multiple SSL engines, or because there is no SSL
version (say when compiled only with Schannel). This change avoids
crashes in that case by returning an empty string for the SSL version
name.

This name is used to determine if SSL requires locking. I think an empty
string works correctly in that case. Only a limited number of *known*
SSL versions do require the locks.