feat(common): create promises without shared state
Sometimes we need to create a `promise<T>` without any shared state,
typically because we won't be able to create the cancellation callback
until later in the code. One can create a promise and replace its shared
state, but that seems wasteful, like creating a `unique_ptr<>` with an
object only to reset it a few lines later. This PR introduces
`google::cloud::null_promise_t`, analogous to `nullopt_t` to initialize
such promises.