ci: remove unneeded explicit build step (#5066)
I don't think there's any need to do an explicit bazel build. Any code and tests will be compiled when they need to be run. Additionally, if a bazel test invocation uses slightly different flags than the bazel build invocation, the cache will not be hit and we'll unnecessarily compile stuff in the bazel build step.