fix(storage): disable XML with custom endpoints
The library automatically uses XML for some operations (for performance
reasons). This does not work well when using a custom endpoint, as the
endpoints for XML are not set.  We completely disable XML if the
application sets a custom endpoint, as in this case they (most likely)
want to use an emulator.