feat(testbench): better emulation for XML requests (#5083)
Because I (coryan) did not know about host-based routing the XML
emulation has used a different subdirectory `/xmlapi/`. The production
service uses the `Host: ` header to resolve the bucket for XML requests,
with this change the testbench does the same.

The old `/xmlapi/` path is still working, I will remove it once the
tests pass without it.