fix(storage): consistent endpoint computation (#5095)
The interaction of `ClientOptions::set_endpoint()` and the
`CLOUD_STORAGE_TESTBENCH_ENDPOINT` environment variable was inconsistent
across endpoints. For JSON endpoints `set_endpoint()` overrode the
`CLOUD_STORAGE_TESTBENCH_ENDPOINT` value, while for XML endpoints it was
the opposite.

In other libraries the environment variable always wins, so we are
changing the behavior here. This behavior was never documented, and it
was buggy, this is not a breaking change per-se, but should be
highlighted in the CHANGELOG.

I refactored the code to compute the different endpoints to standalone
functions. That makes them easier to test.

I used friend functions in `internal::` over member functions in
`ClientOptions` because we will (some day) remove the XML endpoints and
I do not want to make `XmlFoo()` part of the public interface.