feat(storage): disable XML via environment variable (#5100)
When using emulators that do not support XML, application developers may
need to switch off the automatic XML switching in the client library. We
use an environment variable over a `ClientOptions` setting because (a)
this is an implementation detail that we want to change if needed, a
setting would require API changes, and (b) this appears to be needed
only for development / testing, presumably the application developers
don't want to change their code to test against an emulator vs.
production.