cleanup: Add cspell to clang-tidy build (#5107)
Fixes: #5093

This adds spellchecking via cspell to the clang-tidy build. Some code that was
not wrapped in backticks was done so while grammar, punctuation and
capitalization (i.e, grpc to gRPC) was changed. Anything that I feel like
didn't make sense to change got added to the exclude list. We only spellcheck
the comments and exclude certain Doxygen and clang-tidy directives and
backticks to avoid false positives.