test(storage): remove stale buckets (#5160)
Remove stale buckets created by the integration tests and/or the
benchmarks. In both cases I picked one of the integration tests
(`bucket_integration_test`) and one of the benchmarks
(`storage_file_transfer_benchmark`) to perform the cleanup. Doing this
in each test or benchmark is just duplication of work and asking for
"thundering herd"-type problems.