feat(pubsub): wait for callback return *and* handler (#5161)
When shutting down, a subscription session should wait for the callbacks
to return, in addition to waiting for the `AckHandler` to be used (i.e.
either `ack()` or `nack()` gettting called). Without this change it is
very easy to use the callbacks incorrectly, as calling
`std::move(h).ack()` eary in the function may start the shutdown process
and delete objects used later in the function.