cleanup: quiet MSVC+x86 in a single place (#5163)
Including "absl/strings/str_join.h" triggers a warning (and we have
warnings-as-errors enabled). The warning is useful, we don't want to
quiet it everywhere, we just want to silence it for that abseil header.
Instead of sprinkling the directives in each include, let's just wrap
the include with one that also silences things. Once Abseil is quiet we
can fix this with a single search+replace+reformat.