cleanup(doc): provide up-to-date instructions in contributing docs
Fixes: #5236

This fixes a few typos, makes the docs more consistent in some places and
provides up-to-date instructions for installing development tools. I removed
the manual instructions for formatting and linked to `check-style.sh` because I
think that will provide less problems and less updating in the future.  I also
reference Clang 10 as that seems to be the latest version on Fedora, which
also happens to be available in Bionic.