test: enable all clang-tidy presubmit warnings for source files
As discussed in #4230, running clang-tidy on a subset of files produces
some false positives in header files. However, the fix to that also
disabled some warnings for source files. This PR re-enables our full
set of warnings for source files, allowing presubmit to catch more issues
(see #5240 as an example of a break caused by this.)