fix(pubsub): handle errors on first Read() (#5292)
For a streaming pull, the service may not be able to report some errors,
e.g, whether the caller has the right permissions, or whether the
subscription exists, until the first `Read()`. This change modifies the
startup sequence to include the first read. Inlining all of this in a
single lambda seemed too ugly, so we unroll into several callbacks.