feat(pubsub): throughput benchmark tweaks
Much better flow control for the publisher. It is very easy to consume
all the memory on the benchmark application with handlers for responses
and pending RPCs. We need to limit how many responses are pending and
limit how fast we generate requests.

Support setting the endpoint, regional endpoints may be configured with
additional quota for some projects.

Support setting the number of gRPC channels in the publisher and
subscriber.

Split the functions to print the benchmark configuration.