fix(pubsub): missed timers for corked batches
`BatchingPublisherConnection` was (properly) ignoring flush timers when
a batch was corked, i.e., the publisher requires ordering *and* the
previous `AsyncPush()` was pending. However, it also failed to reset the
timer when the batch became uncorked, which can be a problem if (a) the
batch is not large enough to trigger flushing by size *and* (b) that was
the last message sent.