feat(pubsub): handle AsyncPull responses in batches (#5441)
This is more efficient that dispatching each response back through the
completion queue. Since batches can be large (1,000 elements is not
usual) this can save thousands of I/O operations. With this change the
benchmark can sustain 250 MB/s using small (~1KiB) messages, and the
benchmark uses about 5 cores to do so. That is about 5 times better than
before.