feat(pubsub): show ack throughput in benchmark (#5439)
Report (separately) the throughput for queueing messages and processing
acks. This can show when the acks are not being processed fast enough,
or when the messages are queued.

I refactored the code because it was hard to follow what was going on
with all the lambdas. I reduced contention in the benchmark by tracking
pending messages for each sending thread separately.