doc(styleguide): remove sections obsoleted by GSG
Recent (months) updates to the
[GSG](https://google.github.io/styleguide/cppguide.html) have made
several of our differences unnecessary to note, because these
differences are now allowed by the GSG. I've removed these unneeded
sections. Specifically:

* Non-const reference arguments are now allowed and encouraged for
in/out params.
* Preincrement is explicitly preferred unless the postincrement
semantics are explicitly needed.
* The style guide prefers enumerator names `kLikeThis`, and we also
prefer that and it's enforced by our .clang-tidy config. Out styleguide
was just out of date on this point.