fix(pubsub): treat some kUnavailable as success
The service closes streaming pull requests with `kUnavailable` if there
is no data to pull. We should not treat these as failures (not even as
transient failures), because we showed that we successfully connected to
the service, that the subscription exists, and that we have permissions
to pull data. The "problem" is that we do not have data yet.