feat: introduce a generic StreamRange<T> (#5532)
Part of #5521

This PR is part of the work needed for resumable streaming RPCs. This PR introduces a generic StreamRange<T>, which provides an input range of T and an iterator class for iterating it. This is a better, and reusable, version of the RowStreamIterator and TupleStreamIterator classes in spanner link. Indeed, those classes could easily be replaced by aliases to StreamRange<T>, except for small API breaks that would result, because those Spanner classes have weird constructors that this one doesn't have. We can discuss whether we want to make that change or not later.