refactor!: change gcs to use common PaginationRange<T> (#5545)
BREAKING CHANGE

The common PaginationRange<T> class lacks a constructor that existed in the custom range that's being replaced. Specifically, these new ranges will not be constructible with a constructor like:

  explicit PaginationRange(	
      Request request,	
      std::function<StatusOr<Response>(Request const& r)> loader,	
      std::function<std::vector<T>(Response r)> get_items =	
          [](Response r) { ... })
But this constructor was never intended to be part of the public API of these ranges. And in order for a caller to use it correctly they would likely need to reference Request and Response types that are in storage's internal namespace. But nonetheless, this appears to be a potentially breaking change.

I also removed IteratorCompare tests that now fail in cases where they were asserting behavior that isn't required of input iterators. In particular when an input iterator is incremented, all previous copies of the iterator may be (and often are) invalidated. Also, it's UB to compare iterators from different ranges. With these assertions removed, the whole IteratorCompare test was not doing anything unique, so I removed them.