refactor: make StreamRange<T> public
`StreamRange<T>` is a user-facing type that our public APIs return to
callers. Users must be able to read code/documentation to learn how to
use this type. Therefore, `StreamRange<T>` should not live in an
`internal` namespace, otherwise, we're forcing our users to reach into
our internal namespace to use a type that we give them.