feat(spanner): Point-In-Time Recovery (lite) throttled DB update (#5560)
The `UpdateDatabaseDdlMetadata.throttled` field indicates that the
`UpdateDatabase()` operation is throttled, e.g., due to resource
constraints. When resources become available the operation will
resume and the field will be false again.

Here, as part of the `DatabaseAdminClientTest.DatabaseBasicCRUD`
test, we set a valid `version_retention_period` through an
`UpdateDatabase()` statement, and check that the throttled field
was returned correctly (as false).

No samples are required for this field.

Fixes #5257.