feat(spanner): Point-In-Time Recovery (lite) throttled DB update (#5562)
Update the "com_google_googleapis" package to a recent snapshot.
This includes the `UpdateDatabaseDdlMetadata.throttled` field, and
requires adding `table_reference.proto` to the cmake lists.

The `UpdateDatabaseDdlMetadata.throttled` field indicates that the
`UpdateDatabase()` operation is throttled, e.g., due to resource
constraints. When resources become available the operation will
resume and the field will be false again.

This only updates an integration test. No samples are required for
this field.