chore: upgrade grpc to 1.34.0
Also updates protobuf to 3.14.0, because this gRPC required a newer
protobuf. This gRPC also takes a dep on RE2, which is available as
package on many systems, but on some older distros we need to compile it
from source.

Note: This upgrade of grpc will enable us to upgrade our dep on Abseil
to the Sept 2020 LTS, which I'll do in a followup PR.