fix: only build gRPC testing utilities if needed (#5594)
Our customers can enable only specific libraries, for example, they can
enable only the `storage` library. As the storage library does not
require gRPC, we should stop downloading and installing the gRPC
dependencies too (this is one of the main motivations to disable the
other libraries). In this PR we make 3 changes to make this easier to
use:

- Do not build the testing support libraries for gRPC unless gRPC is
  needed by one of the top-level libraries.

- Do not build the storage benchmarks if gRPC is disabled, as they
  depend directly on gRPC.

- Do not install the testing support libraries. These libraries are not
  part of the public API, so installing them is not needed (it used to be,
  to support `google-cloud-cpp-spanner` and other split repos). The other
  changes are easier to implement if installation is disabled.