feat(spanner): introduce spanner::RequestPriority enum (#5624)
Introduce the `spanner::RequestPriority` enum (with low, medium
and high values), and add `absl::Optional<RequestPriority>` members
to `QueryOptions`, `ReadOptions`, and `CommitOptions`. These option
fields can be set by the user to mark a variety of `spanner::Client`
calls with a lower-than-default priority.

See the `RequestOptions.priority` Spanner protobuf field, which will
be set in an upcoming change.