fix: don't wait for state change on shut down CQ
This fixes #5670.

Before this PR, `NotifyOnStateChange()` could be called on a
`grpc::CompletionQueue` which was shut down via `Shutdown()`. This PR
makes `AsyncConnectionReadyFuture` use
`CompletionQueueImpl::StartOperation` to make sure that the whole
operation either fails with `StatusCode::kCancelled` or there is a
guarantee that the `CompletionQueue` is not shut down.