cleanup(firestore): consistent target names
Changed the firestore targets to have the `google-cloud-cpp::` prefix.
Add backwards compatibility CMake and pkg-config files.

Because this library is marked as experimental I named the target
`google-cloud-cpp::experimental-firestore`. That makes it obvious what
its current state is, and we can rename that target later (maybe with a
time limited alias too).