ci: update vcpkg version
In https://github.com/googleapis/google-cloud-cpp/pull/5791 we upgraded
grpc and Abseil, and then in
https://github.com/googleapis/google-cloud-cpp/pull/5799 we changed some
code to rely on new Abseil behavior. This broke tests on Windows because
they were pinned to an older version of vcpkg and therefore and older
version of Abseil.

This PR bumps our vcpkg snapshot to its current head. This should pull
in a newer version of Abseil and grpc that should make our test pass.
However, the newer versions are still not quite as new as we use in our
other builds, because vcpkg does not have the absolute newest releases
of grpc and Abseil. This should be fine.