test: detect OS using standard Bazel configs
Use the standard Bazel configs in `@bazel_tools//src/conditions/...` to
determine what is the target OS and CPU, and use that to determine if
`getrusage()` exists and it supports `RUSAGE_THREAD`. This seems
"better" than our custom config settings, but I like the CMake approach
better (with the know downside of cross-compilation).