feat: make StreamRange<T> public (#5811)
This class is moving from google/cloud/internal/ to google/cloud/, and also in a non-internal namespace. This is now a user-facing class that we want users to be able to use and read the documentation for. We expect to have user-facing functions directly returning objects of type google::cloud::StreamRange<Foo>, so it needs to be public.