ci: only cache certain .cache/ dirs (#5822)
Rather than caching all of ~/.cache, we only cache certain subdirs under
there. This will effectively STOP caching ~/.cache/bazel. But from the
rationale in #5819,
that may be a good thing.

NOTE: This PR may make the bazel builds SLOWER, so we may end up reverting this PR. That's fine.
We won't know if this makes things slower until after this commit makes it through a full CI build where we upload
new cache files that no longer contain the ~/.cache/bazel dir. Then the following builds will NOT be using it.