doc: cross-reference building guides
We had different audiences in mind for our "how to build the code" guides, but
so far we have not been explicit about the intended audience, nor had we made
it particularly easy to find the right document.

This change starts each guide by a "who is this intended for", and then has
links to the other relevant documents in case the reader lands in the wrong one
for some reason.