refactor(spanner): spanner::Timestamp/protobuf::Timestamp conversions
`spanner::Timestamp` and `protobuf::Timestamp` are fundamental parts
of the Spanner admin API (and even more so with the introduction of
PITR), but there was no mechanism provided to the user for converting
between them.

So, `internal::TimestampFromProto()` and `internal::TimestampToProto()`
are here refactored as `spanner::MakeTimestamp(protobuf::Timestamp)`
and `spanner::Timestamp::get<protobuf::Timestamp>()` respectively.

This means Spanner no longer depends on the "internal/time_utils.h"
grab bag.

This also reverses a decision made in #5844 concerning the impossibility
of failure in `protobuf::Timestamp` to `spanner::Timestamp` conversions.
While failure is impossible for `protobuf::Timestamp` values that abide
by the "MUST" comments, in general we must be prepared to handle values
that are out of range.