fix: support for static builds with MSVC
My previous cleanup was incorrect. CMake protects usage of
CMAKE_MSVC_RUNTIME_LIBRARY with a policy, the variable (or rather the
MSVC_RUNTIME_LIBRARY property set by this variable) has no effect if the
CMP0091 policy is set to OLD, as would be in our builds because we say
that version 3.5 is the behavior we want in cmake_minimum_required().