feat: ConnectionOptions<T> can convert to `internal::Options` (#5864)
Part of https://github.com/googleapis/google-cloud-cpp/issues/5738

This give us and users a path to migrate from the old `ConnectionOptions<T>` instances so the new `google::cloud::Options` class. After this PR, we can start adding overloads to existing functions that take `ConnectionOptions<T>` to also take `Options`, and have the `Options` overload be preferred.

## Attention

Since this is the first PR where we're really _using_ the `Options` class, I'll make a few observations and ask for feedback.

1. I don't love that `opts.set<GrpcUserAgentPrefixOption>(std::set<std::string>{user_agent_prefix_});` requires the user to spell the `std::set` part. The reason for this is because `Options::set<T>()` takes a variadic forwarding reference, and passing _only_ an initializer list (e.g., `opts.set<GrpcUserAgentPrefixOption>({user_agent_prefix_});`) will not be able to deduce `U`. We _could_ "fix" this if `Options` takes a position that all option structs _must_ have one data member named `.value` (which we're currently agnostic about). In this case, we could use the type of the `.value` member to _be_ the type of set's argument, thus avoiding the forwarding reference, and thus enabling caller-side initializer lists. Thoughts?
2. `Options` only "gets" _values_, so that requires a copy. This is fine and easy, but it could become an issue if these options become large and get copied a lot. For example, consider an option whose value is a set/map, and the caller wants to insert a value. That would require making a copy, inserting the value, then moving the new value over the old one. Just something to consider.
3. The endpoint option is gRPC specific: `GrpcEndpointOption`. Is that fine? I think it's OK. If we make it transport agnostic, then we'll need to find another header for it to live in.
4. `GrpcUserAgentPrefixOption` has a _std::set_ for all the prefixes. This means that the user can't control the order of the prefixes. Is this OK with us? I think it's probably fine. I don't know if/why the order would matter. If we do care about the order, we could use a `std::deque`.

Note: I did not change the implementation of `ConnectionOptions<T>` to simply hold a single `google::cloud::Options` instance and re-implement its getters/setters in terms of `Options`, because there were enough little semantic differences that made it messy. For example, `channel_pool_domain()` is now implemented as a single key in `GrpcChannelArgumentsOption`, which makes it very difficult for the `channel_pool_domain()` getter to return a const-reference (to what)? I could stuff a default empty string in there for that option, but then an empty channel_pool_domain is used as a signal to not set that channel attribute.