refactor(spanner): spanner::Timestamp/protobuf::Timestamp conversions… (#5876)
`spanner::Timestamp` and `protobuf::Timestamp` are fundamental parts
of the Spanner admin API (and will be even more so with the introduction
of PITR), but there was no mechanism provided to the user for converting
between them.

So, `internal::TimestampFromProto()` and `internal::TimestampToProto()`
are here refactored as `spanner::MakeTimestamp(protobuf::Timestamp)`
and `spanner::Timestamp::get<protobuf::Timestamp>()` respectively.

This means Spanner no longer depends on the "internal/time_utils.h"
grab bag. Time arithmetic is done using `absl::Time`, and expectations
in tests are checked using `spanner::Timestamp`.