ci: prefer curl which is avail by default more often
Curl seems to be more commonly available on default installs of
operating systems, including macOS, so it's nice to prefer curl rather
than wget so we can avoid the user having to install wget when we could
have used curl.

Note: most of our CI scripts were already depending on curl, so this
shouldn't introduce a new dep, rather I think it should remove a dep.