test(spanner): support filtering of `PickRandomInstance()` candidates
Support filtering of `PickRandomInstance()` candidates (1) on the server
side, by exposing the `ListInstances` filter, and (2) on the client side,
via a predicate over `Instance` and `InstanceConfig`.

This will be needed to support the upcoming CMEK integration tests (at
least in their current configuration), where the selected instance must
be in the single location we use to store the encryption keys.

Also use server-side filtering to implement the restriction of instance
IDs to "test-instance-*".

Fixes #5588.