test(spanner): extend samples and integration-test timeouts
We now run the GOOGLE_CLOUD_CPP_SPANNER_SLOW_INTEGRATION_TESTS
overnight, with a couple of new ones either recently submitted
or in the works, so extend the timeout.

The change from 15m to 60m could cause a failure delay if you
hit a hang without even running the slow tests, but so be it.