cleanup: add testing_util::ScopedLog (#5936)
This is basically a replacement for the old testing_util::CaptureLogLinesBackend class that's a bit easier to use because it's an RAII object. I also updated all call sites to use this new class, and then renamed the header and impl file because now all call sites only used the new ScopedLog form.