refactor: add Options::get<T>()
It turns out that we actually can provide a const `get<T>()` method that
returns a const-ref by defining that it returns a reference to a static
value-initialzied default value. I think this will be very useful for
cases where we want to query an option with changing or setting it. For
example, in a number of places we inspect the `TracingComponentsOption`
which is a `std::set<std::string>` to see if it contains an element.
Using `.get<T>()` would let us do this without copying the `set` every
time. I believe this will also allow us to re-implement
`ConnectionOptions` in terms of `Options`, since it can now return
const-refs, even to unset options.