test(spanner): matching test instance with test resources (#5956)
Undo the addition of a predicate to `PickRandomInstance()` (from
#5885) to, say, one in a specific location where test resources were
configured.

A better tactic is to configure test resources in all possible
test-instance locations.  So, introduce `InstanceLocation()`, which
can be used to select location-specific test resources for the
randomly selected instance.

This will be needed to support the upcoming CMEK integration tests
and samples, where the location of the encryption keys must match
the location of the instance.