fix(storage): do not update upload session ids
Our documentation said the upload session id could change. This does not
seem to happen in production, and the APIs are harder to use if we
accept changes.  I modified to the doxygen documentation, and prevented
changes from having any effect (even if the changes are not supposed to
happen).