refactor(spanner): add spanner_internal::DefaultOptions() in preparation for using `Options` (#5985)
Related to #5738
This PR breaks off a piece of: #5984

This PR adds the spanner_internal::DefaultOptions() function (google/cloud/spanner/internal/options.h). All the defaults for spanner moved into the body of this function. The existing ConnectionOptionsTraits class was changed to pull its defaults out of spanner_internal::DefaultOptions(), so that we don't have defaults living in two places.