refactor(spanner): remove unnecessary MakeConnection factory (#5995)
There's no need for a factory to construct
spanner_internal::ConnectionImpl. We can just construct it directly
from client.cc. I don't think the factory was adding any value, but it
was adding an extra layer of things to plumb parameters through.
This also makes the code easier to follow.