test(common): deflake tests for future<T>
The tests to verify `.wait_for()` actually blocks for a bit were flaky,
and overly complicated. The new version just measures the time spent in
the `.wait_for()` operation, if it is not-zero milliseconds we declare
success. A more sophisticated test would measure the wait time with
different sleep times, and run some statistics to convince ourselves the
implementation is "correct". We are relying on knowledge of how this is
implemented, and trusting the C++ library, maybe bad assumptions, but
reasonable tradeoffs for faster tests.