cleanup(spanner): new version time in spanner_create_backup sample (#6016)
A newly changed requirement on the PITR `CreateBackup()` sample
(see #5943 for the previous change) to hide the source of the
`version_time` from the snippet altogether by passing it as a
parameter.  Let's extend that decision to `expire_time` too.

In the process, let's avoid cross-clock-domain issues as well,
by only using timestamps originated by the service.  In these
cases we only need "now", from `SELECT CURRENT_TIMESTAMP()`
instead of from `absl::Now()`.