refactor(spanner): add spanner_internal::MakeConnection (#6022)
This is the function that will be the preferred user-facing
"MakeConnection" function. It's requires a spanner::Database argument,
and nothing else. A single, optional Options argument may be specified
to configure aspects of the returned Connection. Appropriate defaults
will be used for any unspecified option, so {} is a valid (and the
default) argument.

The other MakeConnection overloads will continue to exist and they
will simply forward to this one new/preferred overload. The old
overloads will NOT be deprecated, so we will not break existing callers.
However, we will clearly document all the overloads to make it clear
that the new overload should be preferred, so that customers do not
struggle to decide which overload to choose.

Of course, this is all "internal" for now, but it will likely be moved
out of internal and made public soon (before the April release).