refactor: rework the Status/StatusOr<T> gmock matchers (#6025)
Because `StatusOr<T>` does not have a printer (and should not), gmock
will render its value as "N-byte object <...>", which is not very
useful. Accordingly, `StatusIs` will now print the enclosed `Status`
directly so that a failing expectation will not require further
explanation.

This means that we no longer need workarounds like ...
```
  EXPECT_THAT(v, IsOk()) << v.status();
```
or
```
  EXPECT_THAT(v.status(), IsOk())
```
just to be able to see the entire `Status<T>::status()` in a failed
expectation.

It also means we don't have to, nor want to, print a `Status` code
or message during either of their sub-matches.

Also generate an explanation even when a `StatusIs` match succeeds.
This is necessary when the `StatusIs` matcher is used as an argument
to another matcher that can change a success into a failure (e.g.,
`Not()`).  This means that ...
```
  EXPECT_THAT(v, Not(StatusIs(StatusCode::kInternal)));
```
will now describe the failing internal Status.

Finally, implement `EXPECT_STATUS_OK` using `StatusIs`. This means the
expectation output is the same for all matcher varieties.  The helper
macros are now exported by `google/cloud/testing_util/status_matchers.h`,
although `assert_ok.h` remains for backwards compatibility.