refactor(spanner): db admin connection uses new Options (#6026)
This PR changes adds a MakeDatabaseAdminConnection overload (currently
in spanner_internal) that accepts the new Options. This is the
overload that will be the recommended user-facing factory once
Options become public. All other factories forward to this one (except
for the one overload that exists only for testing.)

This PR also adds spanner_internal::DefaultAdminOptions, which will create
the correct defaults for Spanner's admin operations.