cleanup(spanner): remove some TODOs
Cleans up some TODOs related to
https://github.com/googleapis/google-cloud-cpp/issues/5738.

Some of the TODOs are able to be completed now.

The TODO in connection_option.cc was not done, but I decided that it
doesn't actually need to be done. Caching the endpoint in a
function-local static is purely a performance optimization, which we
could do at any point if it's an _actual_ performance issue in practice.
Until then, it's safest to preserve backward compatible semantics.