feat(common): make Options public (#6042)
Part of #5738

I think we're ready for this now. This class is now used in the google/cloud/spanner/ code to implement all the options passing, and the API has been holding up well and seems simple to use and read. This class is also already used to provide several different types of options, such as the common options, the gRPC options, and two sets of Spanner-specific options (spanner policies and session pool options). At this point, I'm quite confident that it will work nicely for all of our options needs.

This PR is a mechanical renaming. No semantic changes. There are a couple comment edits to make things read a bit nicer.

Planned follow ups from this PR:

Update https://github.com/googleapis/google-cloud-cpp/blob/master/ARCHITECTURE.md to explain how Options are intended to work and be used.
Move the Spanner functions that use Options out from spanner_internal:: -> spanner::
Unblock the Generator work so that it can use the newly public Options code (#6019)