cleanup(bigtable)!: make *Client::BackgroundThreadsFactory() pure (#6047)
Change the recently-added BackgroundThreadsFactory() functions (see
and `bigtable::InstanceAdminClient to be pure virtual.

This restores these class to be pure interfaces, rather than having one
small piece of concrete behavior.  A default implementation that returns
a null factory, which throws when called, is never useful.

This may seem like a breaking change, but it was actually the introduction
of the member function that was breaking, as derived classes already had
to override the function in order to use the client (for example, to pass
a `DataClient` to the `bigtable::Table` constructor).

Failure to override will at least now be a compilation error, rather than
a runtime error (`bad_function_call`).

This was discovered while trying to use the latest Bigtable client in
another package, and seeing exceptions thrown.