ci: require sourced modules to specify full path (#6054)
Part of https://github.com/googleapis/google-cloud-cpp/issues/3911

This change requires that shell libraries loaded with `source module
<file>` must specify the file via a full path starting from the project
root. This makes it easier for humans to see sourced files and know
which file on disk is actually loaded. This also more closely matches
the full paths we use in C++ includes.

Note the leading `/` in these sourced paths is optional. I'm including
it because I thought it made it look clear that the path was "absolute"
(well, relative to the project root).